feat(CustomerUserDetailsService): Add CustomUserDetailsService for user authentication
Introduced the CustomUserDetailsService class, which implements the UserDetailsService interface provided by Spring Security. The class is annotated with @Service to indicate that it is a service component. The CustomUserDetailsService is responsible for loading user details by their username from the UserRepository. If a user is not found, a UsernameNotFoundException is thrown. This service class enables user authentication and retrieval of user details for authentication purposes.